+++ packages/axios-extension/src/abap/adt-catalog/services/index.ts
@@ -7,3 +7,4 @@ export { FileStoreService } from './filestore-service';
 export { GeneratorService } from './generator-service';
 export { BusinessObjectsService } from './businessobjects-service';
 export { UI5RtVersionService } from './ui5-rt-version-service';
+export { AbapCDSViewService } from './abapcdsview-service';