+++ packages/create/src/cli/change/change-data-source.ts
@@ -1,12 +1,15 @@
 import type { Command } from 'commander';
-import type { AdpPreviewConfig, DescriptorVariant, PromptDefaults } from '@sap-ux/adp-tooling';
-import { generateChange, ChangeType, getPromptsForChangeDataSource, getManifest } from '@sap-ux/adp-tooling';
+import {
+    generateChange,
+    ChangeType,
+    getPromptsForChangeDataSource,
+    getAdpConfig,
+    getManifestDataSources,
+    getVariant
+} from '@sap-ux/adp-tooling';
 import { getLogger, traceChanges } from '../../tracing';
-import { readFileSync, existsSync } from 'fs';
-import { join, isAbsolute } from 'path';
-import { UI5Config } from '@sap-ux/ui5-config';
 import { promptYUIQuestions } from '../../common';
-import { getAppType } from '@sap-ux/project-access';
+import { validateAdpProject } from '../../validation/validation';
 
 let loginAttempts = 3;
 
@@ -20,7 +23,7 @@ export function addChangeDataSourceCommand(cmd: Command): void {
         .option('-s, --simulate', 'simulate only do not write or install')
         .option('-c, --config <string>', 'Path to project configuration file in YAML format', 'ui5.yaml')
         .action(async (path, options) => {
-            await changeDataSource(path, { ...options }, !!options.simulate, options.config);
+            await changeDataSource(path, !!options.simulate, options.config);
         });
 }
 
@@ -28,41 +31,19 @@ export function addChangeDataSourceCommand(cmd: Command): void {
  * Changes the data source of an adaptation project.
  *
  * @param {string} basePath - The path to the adaptation project.
- * @param {PromptDefaults} defaults - The default values for the prompts.
  * @param {boolean} simulate - If set to true, then no files will be written to the filesystem.
  * @param {string} yamlPath - The path to the project configuration file in YAML format.
  */
-async function changeDataSource(
-    basePath: string,
-    defaults: PromptDefaults,
-    simulate: boolean,
-    yamlPath: string
-): Promise<void> {
+async function changeDataSource(basePath: string, simulate: boolean, yamlPath: string): Promise<void> {
     const logger = getLogger();
     try {
         if (!basePath) {
             basePath = process.cwd();
         }
-        if ((await getAppType(basePath)) !== 'Fiori Adaptation') {
-            throw new Error('This command can only be used for an Adaptation Project');
-        }
-        checkEnvironment(basePath);
-
+        await validateAdpProject(basePath);
         const variant = getVariant(basePath);
-        const ui5ConfigPath = isAbsolute(yamlPath) ? yamlPath : join(basePath, yamlPath);
-        const ui5Conf = await UI5Config.newInstance(readFileSync(ui5ConfigPath, 'utf-8'));
-        const customMiddlerware =
-            ui5Conf.findCustomMiddleware<{ adp: AdpPreviewConfig }>('fiori-tools-preview') ??
-            ui5Conf.findCustomMiddleware<{ adp: AdpPreviewConfig }>('preview-middleware');
-        const adp = customMiddlerware?.configuration?.adp;
-        if (!adp) {
-            throw new Error('No system configuration found in ui5.yaml');
-        }
-        const manifest = await getManifest(variant.reference, adp, logger);
-        const dataSources = manifest['sap.app'].dataSources;
-        if (!dataSources) {
-            throw new Error('No data sources found in the manifest');
-        }
+        const adpConfig = await getAdpConfig(basePath, yamlPath);
+        const dataSources = await getManifestDataSources(variant.reference, adpConfig, logger);
         const answers = await promptYUIQuestions(getPromptsForChangeDataSource(dataSources), false);
 
         const fs = await generateChange<ChangeType.CHANGE_DATA_SOURCE>(basePath, ChangeType.CHANGE_DATA_SOURCE, {
@@ -81,35 +62,9 @@ async function changeDataSource(
         if (error.response?.status === 401 && loginAttempts) {
             loginAttempts--;
             logger.error(`Authentication failed. Please check your credentials. Login attempts left: ${loginAttempts}`);
-            await changeDataSource(basePath, defaults, simulate, yamlPath);
+            await changeDataSource(basePath, simulate, yamlPath);
             return;
         }
         logger.debug(error);
-    }
-}
-
-/**
- * Get the app descriptor variant.
- *
- * @param {string} basePath - The path to the adaptation project.
- * @returns {DescriptorVariant} The app descriptor variant.
- */
-function getVariant(basePath: string): DescriptorVariant {
-    return JSON.parse(readFileSync(join(basePath, 'webapp', 'manifest.appdescr_variant'), 'utf-8'));
-}
-
-/**
- * Check if the project is a CF project.
- *
- * @param {string} basePath - The path to the adaptation project.
- * @throws {Error} If the project is a CF project.
- */
-function checkEnvironment(basePath: string): void {
-    const configJsonPath = join(basePath, '.adp', 'config.json');
-    if (existsSync(configJsonPath)) {
-        const config = JSON.parse(readFileSync(configJsonPath, 'utf-8'));
-        if (config.environment === 'CF') {
-            throw new Error('Changing data source is not supported for CF projects.');
-        }
     }
 }